Generate zlabels for changes in verbatim environments
This commit adds a new command to mark beginning and end of a segment
that changed. This allows to mark changes in verbatim environments. The
difference from \DIFaddbegin &Co is that there is no further semantics
about changes given by this tag. For undo the diff, this command is simply
to be ignored.